requirements.txt
# ======================================
# Sales Agent MVP - Pinned Dependencies
# ======================================
# Generated from requirements.in
# DO NOT edit manually - regenerate with pip-compile

# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
uvloop==0.21.0
pydantic==2.9.2
pydantic-settings==2.6.1

# ASR
realtimestt==0.3.5
faster-whisper==1.0.3
torch==2.5.1
torchaudio==2.5.1

# VAD
silero-vad==5.1

# LLM Client (OpenAI-compatible для vLLM)
openai==1.54.0

# TTS
piper-tts==1.2.0
onnxruntime-gpu==1.20.1

# Policy & FSM
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.9

# Google Sheets
gspread==6.1.4
google-auth==2.36.0
gspread-asyncio==1.9.2

# Network & Async
aiohttp==3.11.7
websockets==13.1
httpx==0.28.0

# Utils
tenacity==9.0.0
redis==5.2.0
python-dotenv==1.0.1
pyyaml==6.0.2

# Monitoring
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-exporter-otlp==1.28.2

# Audio Processing
soundfile==0.12.1
numpy==2.1.3
librosa==0.10.2.post1

# System Monitoring
psutil==6.1.0
gputil==1.4.0

# vLLM (для download_models.py)
huggingface-hub==0.26.2

# Dev/Test
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
ruff==0.8.0
black==24.10.0
mypy==1.13.0

# Type stubs
types-pyyaml==6.0.12.20240917
types-redis==4.6.0.20241004

